Explicito generación de ints o floats random según el caso

+++ Server/src/Model/Defence.cpp
@@ -1,8 +1,9 @@
 #include "../../includes/Model/Defence.h"
+
 #include "../../includes/Model/Character.h"  // Evito circular dependeces.
 //------------------------------------------------------------------------
-#include "../../../Common/includes/RandomNumberGenerator.h"
 #include "../../../Common/includes/Exceptions/Exception.h"
+#include "../../../Common/includes/RandomNumberGenerator.h"
 
 Defence::Defence(const DefenceCfg& data)
     : Wearable(data.id, data.name, data.price, data.type),
@@ -12,7 +13,8 @@ Defence::~Defence() {}
 
 const unsigned int Defence::use(Character& user) {
     RandomNumberGenerator random_number_generator;
-    return random_number_generator(min_defence, max_defence);
+    return (int)random_number_generator((int)min_defence,
+                                          (int)max_defence);
 }
 
 const unsigned int Defence::getRange() const {
@@ -20,5 +22,6 @@ const unsigned int Defence::getRange() const {
 }
 
 const bool Defence::isHealing() const {
-    throw Exception("Defence::isHealing: Las defensas no pueden ser curativas.");
-} 
+    throw Exception(
+        "Defence::isHealing: Las defensas no pueden ser curativas.");
+}
